mcp-client/pyproject.toml
[project]
name = "tfl-mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.57.1",
    "mcp>=1.11.0",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
]

[dependency-groups]
dev = [
    "ruff>=0.12.3",
]

[tool.ruff]
line-length = 120
